+++ src/ninja/WindNinjaInputs.h
@@ -94,7 +94,9 @@ public:
         noInitializationFlag,			//no initialization, used to check if it has been set or not
         domainAverageInitializationFlag,	//single domain-averaged input speed and direction
         pointInitializationFlag,		//single or multiple point inititialization
-        wxModelInitializationFlag	//Weather forecast model initialization
+        wxModelInitializationFlag,	//Weather forecast model initialization
+        griddedInitializationFlag,  //gridded speed and direction
+        foamInitializationFlag
     };
 
     eVegetation vegetation;
@@ -121,6 +123,8 @@ public:
     /*-----------------------------------------------------------------------------
      *  Input speed/direction Parameters
      *-----------------------------------------------------------------------------*/
+    std::string speedInitGridFilename;  //raster file of gridded directions speeds 
+    std::string dirInitGridFilename;  //raster file of gridded wind directions
     eInitializationMethod initializationMethod;	//method to initialize WindNinja
     std::string forecastFilename;	//name of coarse weather model initialization file (NDFD, NAM, GFS, RUC, etc.)
     velocityUnits::eVelocityUnits inputSpeedUnits;			//units of input windspeed (0=>m/s, 1=>mph, 2=>kph) (note that inputSpeed is always stored as m/s, and converted to and from the other units)
@@ -144,11 +148,6 @@ public:
      *  Surface Properties
      *-----------------------------------------------------------------------------*/
     surfProperties surface;		//surface properties data (roughness, albedo, etc...)
-
-    /*-----------------------------------------------------------------------------
-     *  Stability Properties
-     *-----------------------------------------------------------------------------*/
-    //bool stabilityFlag;  //flag specifying if non-neutral stability parameters should be set
 
     /*-----------------------------------------------------------------------------
      *  Diurnal Inputs
@@ -220,7 +219,10 @@ public:
     std::string pdfDEMFileName;
     std::string pdfFile;
     double      pdfResolution;
+    double      pdfLineWidth;
     lengthUnits::eLengthUnits pdfUnits;
+
+    std::string customOutputPath; //user-specified path for output
 
     /*-----------------------------------------------------------------------------
      *  Ninja Speed Testing Section
@@ -262,16 +264,6 @@ public:
     bool stabilityFlag;  //flag specifying if non-neutral stability parameters should be set
 #endif
 
-    /*-----------------------------------------------------------------------------
-     *  SCALAR section
-     *-----------------------------------------------------------------------------*/
-#ifdef SCALAR
-    bool scalarTransportFlag;  //flag specifiying if scalar transport should be calculated
-    double scalarSourceStrength;
-    double scalarSourceXcoord;  //input long
-    double scalarSourceYcoord;  //input lat
-#endif //SCALAR
-
     std::string outputPointsFilename; //name of file containing output for requested point locations
     std::string inputPointsFilename; // name of file containing locations of specfic points for output
     
@@ -293,7 +285,6 @@ public:
     int nIterations; //number of iterations for a ninjafoam simulation
     int meshCount; //mesh count for a ninjafoam simulation
     eNinjafoamMeshChoice ninjafoamMeshChoice; // fine, medium, coarse
-    eMeshType meshType; //MDM (moveDynamicMesh), SHM (snappyHexMesh)
     bool nonEqBc; //flag indicating if non-equilbrium boundary conditions should be used for a ninjafoam simulation
     std::string stlFile; //path/filename of an STL file
 #endif